requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.17
pymysql==1.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.26.151
alembic==1.11.1
redis==4.5.5
celery==5.3.1
pydantic==1.10.11
